Meta+Editor: Update submodule, introduce tool system.
We can now switch beween the Selection tool (default), and the Place
Object tool. Objects are a group of tiles that make up a bigger object.
I've added two to start with, being a Life Crystal and the Work Bench.
The work bench has multiple styles, as it can be crafted with different
kinds of wood -- this can be modified before placing the object.